Adds the concept of iterations of ticks.
In order to fix a bug regarding integer precision in BYOND,
a concept of iterations of of 2 ^ 24 ticks has been added. The
scheduler's internal counter rolls over at 2 ^ 24, and events
intended for a tick count over 2 ^ 24 is scheduled for 2 ^ 24,
with a remainder stored.

Due to the integer precision in BYOND, values given above 2 ^ 24
will be in floating point representation, and the library will only
be able to schedule them to +/- 1 tick of the intended value.